chore: refactor resolve utils in a separate file
